+++ src/domain/challenge/base-challenge/base.challenge.module.ts
@@ -5,7 +5,7 @@ import { CommunityModule } from '@domain/community/community/community.module';
 import { ContextModule } from '@domain/context/context/context.module';
 import { Module } from '@nestjs/common';
 import { AuthorizationModule } from '@core/authorization/authorization.module';
-import { NamingModule } from '@src/services/domain/naming/naming.module';
+import { NamingModule } from '@services/infrastructure/naming/naming.module';
 import { BaseChallengeService } from './base.challenge.service';
 import { BaseChallengeAuthorizationService } from './base.challenge.service.authorization';
 import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/authorization.policy.module';
