+++ src/domain/challenge/hub/dto/hub.args.query.hubs.ts
@@ -1,6 +1,6 @@
 import { IDsQueryArgs } from '@domain/common/query-args/ids.query.args';
 import { ArgsType, Field } from '@nestjs/graphql';
-import { HubFilterInput } from '@services/domain/hub-filter/dto/hub.filter.dto.input';
+import { HubFilterInput } from '@services/infrastructure/hub-filter/dto/hub.filter.dto.input';
 
 @ArgsType()
 export class HubsQueryArgs extends IDsQueryArgs {
