+++ src/domain/collaboration/callout/callout.module.ts
@@ -10,7 +10,7 @@ import { CalloutService } from './callout.service';
 import { CalloutAuthorizationService } from './callout.service.authorization';
 import { CalloutResolverFields } from './callout.resolver.fields';
 import { CalloutResolverSubscriptions } from './callout.resolver.subscriptions';
-import { NamingModule } from '@services/domain/naming/naming.module';
+import { NamingModule } from '@services/infrastructure/naming/naming.module';
 import { CommentsModule } from '@domain/communication/comments/comments.module';
 import { ActivityAdapterModule } from '@services/adapters/activity-adapter/activity.adapter.module';
 import { NotificationAdapterModule } from '@services/adapters/notification-adapter/notification.adapter.module';
