+++ src/domain/collaboration/collaboration/collaboration.module.ts
@@ -5,7 +5,7 @@ import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { CalloutModule } from '@domain/collaboration/callout/callout.module';
 import { CollaborationService } from '@domain/collaboration/collaboration/collaboration.service';
 import { Collaboration } from '@domain/collaboration/collaboration/collaboration.entity';
-import { NamingModule } from '@services/domain/naming/naming.module';
+import { NamingModule } from '@services/infrastructure/naming/naming.module';
 import { CollaborationResolverMutations } from '@domain/collaboration/collaboration/collaboration.resolver.mutations';
 import { CollaborationResolverFields } from '@domain/collaboration/collaboration/collaboration.resolver.fields';
 import { RelationModule } from '@domain/collaboration/relation/relation.module';
