+++ src/domain/community/contributor/contributor.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ProfileModule } from '@domain/community/profile/profile.module';
 import { AgentModule } from '@domain/agent/agent/agent.module';
-import { NamingModule } from '@src/services/domain/naming/naming.module';
+import { NamingModule } from '@services/infrastructure/naming/naming.module';
 import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/authorization.policy.module';
 import { KonfigModule } from '@src/platform/configuration/config/config.module';
