+++ src/domain/context/context/context.module.ts
@@ -10,7 +10,7 @@ import { AuthorizationModule } from '@core/authorization/authorization.module';
 import { ContextAuthorizationService } from './context.service.authorization';
 import { AuthorizationPolicyModule } from '@domain/common/authorization-policy/authorization.policy.module';
 import { VisualModule } from '@domain/common/visual/visual.module';
-import { NamingModule } from '@services/domain/naming/naming.module';
+import { NamingModule } from '@services/infrastructure/naming/naming.module';
 import { LocationModule } from '@domain/common/location';
 
 @Module({
