+++ src/services/api/conversion/dto/convert.dto.challenge.to.hub.input.ts
