+++ src/services/api/roles/dto/roles.dto.input.organization.ts
@@ -1,6 +1,6 @@
 import { UUID_NAMEID } from '@domain/common/scalars';
 import { Field, InputType } from '@nestjs/graphql';
-import { HubFilterInput } from '@services/domain/hub-filter/dto/hub.filter.dto.input';
+import { HubFilterInput } from '@services/infrastructure/hub-filter/dto/hub.filter.dto.input';
 
 @InputType()
 export class RolesOrganizationInput {
