+++ src/services/api/roles/dto/roles.dto.input.user.ts
@@ -1,6 +1,6 @@
 import { UUID_NAMEID_EMAIL } from '@domain/common/scalars';
 import { Field, InputType } from '@nestjs/graphql';
-import { HubFilterInput } from '@services/domain/hub-filter/dto/hub.filter.dto.input';
+import { HubFilterInput } from '@services/infrastructure/hub-filter/dto/hub.filter.dto.input';
 
 @InputType()
 export class RolesUserInput {
