+++ src/services/api/roles/roles.module.ts
@@ -15,7 +15,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Opportunity } from '@domain/collaboration/opportunity/opportunity.entity';
 import { Challenge } from '@domain/challenge/challenge/challenge.entity';
 import { PlatformAuthorizationModule } from '@src/platform/authorization/platform.authorization.module';
-import { HubFilterModule } from '@services/domain/hub-filter/hub.filter.module';
+import { HubFilterModule } from '@services/infrastructure/hub-filter/hub.filter.module';
 
 @Module({
   imports: [
